fix: update package-lock.json dependencies for babel vulnerability
Also considered removing package-lock.json,
but it is still useful while collaborating on developing it.

The package-lock.json is not included in the files downloaded from NPM
when included as a dependency anyways. This also means that it should be
unnecessary to publish a new version for this update, so no version bump
will follow.

CVE ID
CVE-2023-45133
GHSA ID
GHSA-67hx-6x53-jw92