refactor: adjust and simply config load flow
Need project directory for a future bug fix and
took the opportunity to also the config load flow
some simplification.

Errors are now thrown and caught when configuration is bad instead
of simply returned. For the sake of simplicity, seems fine to
consider that an exceptional state despite it being largely user input.