fix: avoid matching against non-project path parts for patterns
The pattern code matched against the absolute path to a file,
which would unexpectedly match simple patterns for many users such as
`home`, `usr`, and other common path parts.

We avoid this by calculating the relative path from the project root
to the file being analyzed. The result of `context.getFilename()`
may not necessarily be the relative path from the project root,
as eslint could be executed from locations other than the project root.

closes #11