feat: add `recommended` config
Added the configuration suggested in the primary README.
While there may be some argument that barrel files
should be supported as a part of the README,
the current implementation does not have great support
to limit it only to those exports.

Splitting out the rules would be a good step towards that direction;
however, that would require some more work that will be left to
a future change.

In the mean time, the recommended config is kept simple
to just require all imports be absolute.

Closes #12