feat!: add support for resolving tsconfig from the file path
Given TSConfig and ESLint both seem to support nested
configuration files and resolving the most relevant one
based off the current file, it seems reasonable that this
plugin should also.

The preferred tsconfig will be the tsconfig closest to the file moving
forward with additional resolution based off the current file's directory
as before included. If we only went with the former, find-up
could potentially not resolve something that was previous
resolvable with an explicit path like
```
- package.json
- src
        - foo.ts
- config
        - tsconfig.json
```

Despite the new default behavior, it seems unlikely to produce a result
that would cause linter complaints that would not have existed within
TypeScript itself. Thus, the major version will not be bumped following
this.

Unfortunately, as the functions for resolving configuration are currently
mocked at the moment, there would have to be a bit of investment to add
those tests. Coupled with the fact that there would probably be a
significant rewrite to support #10, and I'm okay leaving this without
tests for the time being.

Closes #15

BREAKING CHANGE: TSConfig file resolution may resolve to a different
file than before, potentially causing new lint violations.