fix: resolve regression crashing ESLint when TSConfig is misconfigured
This was a regression as a result of `dd775af`.
It feels as if the function call for a value without
using the value should have been caught by a lint issue;
however, it doesn't seem like JavaScript/TypeScript has
a way to actually designate functions pure, so it is likely
difficult for any ESLint rule to detect whether a function
is pure.