refactor: use tsconfig-paths@^4 to resolve tsconfig/jsconfig, remove find-up
This removes a direct dependency and reduces the amount of code that
needs to be maintained. `tsconfig-paths` has built in behavior to
resolve `tsconfig.json`/`jsconfig.json` starting from the current directory.
If a directory is passed into the function, it will attempt to find
a file with specifically those two names starting from the directory
given. If a file path is given instead it will attempt to load the file
given. These in combination is essentially the existing behavior.

There may technically be some case where the TSConfig file was resolved
to be some combination of the aliasConfigPath and the ESLint config
directory from before; however, that would be a behavior that should be
considered a bug in my opinion. It seems unlikely to affect any
workspaces, and if it does, those workspaces should be considered
misconfigured.