feat: add option to disallow baseUrl resolved paths
This option forces the user of the lint to use a defined
alias rather than simply resolving an absolute path from
the `baseUrl` defined in TSConfig.

The implementation is a bit haphazard and it could be split up
into a separate rule; however, I didn't want to scope that work
into here. It may be split up into a separate rule with the overall
splitting of rules come version 2.0.0. For now, left it as an option
to the existing rule.

Closes #16