feat: add support for eslint@9
With the help of the example of #18,
these changes should add support for eslint@9.

When building and running with the commit in that
pull request, eslint appears to act up locally.
There's probably some odd interaction with the declaration
of it becoming a ESM module. The module treatment likely
also affected how jest handled the mocking in the tests
that ultimately resulted in test failures.

In the end, it ended up being a bit easier to fix it by
starting clean and porting over necessary changes from the PR.

`rimraf` was mistakenly not declared as a direct dev dependency
before and was being used transiently from jest dependencies.
The dependency updates thus required that we pull it in
such that rimraf still exists as a dependency for us.