build: v1.5.0
Given that the typescript-eslint plugin upgrade
simply added support for eslint@9 (and was thus the
basis for the addition to peer dependency), the support
for eslint@9 should be backwards compatible with older versions.
Thus, only the minor version is being bumped for this change.

If this proves to be incorrect, the version should be deprecated
and a 2.0.0 version should be released instead.