fix: correct eslint@8 dependencies post-support for eslint@9
It appears that older versions of eslint@8 does not have
full forward compatibility support for eslint@9.
Configuration with this plugin would error in the older version
with the old configuration scheme, so we'll need to specify
newer versions. From some emperical testing, it seems like
8.40.0 is sufficient for this.

Not fully certain if this incompatbility exists for even older version
(eslint@2~7) at the moment.