fix: allow aliases which match baseUrl-resolved imports for !isAllowBaseUrlResolvedImport
This is an odd case where if an alias is specified for an
import that is also resolved through an absolute path
by the baseUrl, it is considered an error.

More concretely, this means that if the baseUrl
resolves to some root folder and the source code is in
`src`, if you defined a `src/*` alias that resolves to
`src/*`, the plugin will flag the imports as invalid
despite matching a provided path.