fix: resolve detection of alias usage when using aliases with slashes
The existing implementation had used a split on `'/'` to determine
whether they were currently using an alias config, which fails
when they use an alias with a slash within it.

This changes the implementation to check whether or not their imported
path begins with the actual alias or not, so that we are able to
make an updated recommendation on which alias to use. As pre-aliased
paths are not relative paths and the existing implementation was not
able to determine whether or not they were using an alias, it could
not construct a complete absolute path to the module and thus assumed
that they were valid imports/library imports.

Closes #20