feat: support enforcing relative imports
This adds support for an additional configuration value
`isRelativeImportOverridesEnforced` which allows users
to strictly require relative imports when it is convenient.

While this goes against the benefit of being able to easily
refactor the entire code-base's imports with consistent imports,
it does still have a consistent and logical rule and generally
falls within the purpose of the plugin to force import style.

Closes #19